+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Quest | Add ability to view previous QuestionnaireResponse
 - Quest | Added workflow configurations
 - EIR | Added workflow configurations
+- ANC | Report Measures List Items, Filter Selection
+- Quest | Add photo capture questionnaire widget
 
 ### Fixed
 -